Rename module `compat._2and3` to `compat.py2and3`
