+++ gc3libs/compat/py2and3.py
